+++ src/finchlite/interface/__init__.py
@@ -106,8 +106,8 @@ from .fuse import (
     set_default_scheduler,
 )
 from .lazy import LazyTensor, asarray, lazy
-from .np_wrapper import NumPyFType, NumPyWrapper
-from .scalar import Scalar, ScalarFType
+from ..tensor.np_wrapper import NumPyFType, NumPyWrapper
+from ..tensor.scalar import Scalar, ScalarFType
 
 __all__ = [
     "EagerTensor",
