+++ tests/test_logic_compiler.py
@@ -3,7 +3,7 @@ import numpy as np
 import finchlite.finch_logic as logic
 from finchlite import ffuncs, ftype
 from finchlite.autoschedule import INTERPRET_NOTATION, NotationGenerator
-from finchlite.compile.bufferized_ndarray import (
+from finchlite.tensor.bufferized_ndarray import (
     BufferizedNDArray,
 )
 from finchlite.finch_logic import (
